another try to improve initial frame sizing
Subtract a fixed constant 121 from display height; this works better on
Windows. I measured it as the difference between the outer and inner
frame-edges.
Probably needs tweaking for Mac and Linux.
I'm not sure how to get that 121 value dynamically, because we don't
have any frames already mapped when this is called.