Fixes for latest Emacs on Windows: emacs bin dir & scoop msys root
I'm now using `scoop` as a package manager on Windows, and
scoop-installed msys goes in ~/scoop/apps/msys2/current, so have
to add that to exec-path and $PATH.

Also latest emacs needs to have its own invocation directory in $PATH
to find compiler and linker for native-comp. NOTE: this works, but
beware of overwriting $PATH later when getting system env.